requirements.txt
Django>=4.2
djangorestframework
python-dotenv
django-cors-headers
PyMuPDF
openai
faiss-cpu
numpy
uvicorn
gunicorn
